refactor(server): api key auth
